[GOSC] Update Guest OS Customization to utilize systemd system init
Currently the "telinit 6" command is used to reboot a Linux VM
following Guest OS Customization.  As the classic Linux init system,
SysVinit, is deprecated in favor of a newer init system, systemd,
the telinit command may not be available on the base Linux OS.

This change adds support to Guest OS Customization for the systemd init
system.  If the modern init system, systemd, is available, then a
"systemctl reboot" command will be used to trigger reboot.  Otherwise,
the "telinit 6" command will be used assuming the traditional init
system, SysVinit, is still available.